[hsa] Fix i686 fprintf warning
2016-01-18  Martin Jambor  <mjambor@suse.cz>

	* hsa-dump.c (dump_hsa_symbol): Add missing argumet cast.



git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/hsa@232526 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/hsa-dump.c
@@ -720,7 +720,7 @@ dump_hsa_symbol (FILE *f, hsa_symbol *symbol)
 	   hsa_type_name (symbol->m_type & ~BRIG_TYPE_ARRAY_MASK), name);
 
   if (symbol->m_type & BRIG_TYPE_ARRAY_MASK)
-    fprintf (f, "[%lu]", symbol->m_dim);
+    fprintf (f, "[%lu]", (unsigned long) symbol->m_dim);
 }
 
 /* Dump textual representation of HSA IL operand OP to file F.  */
